The `DragonCode\LaravelFeed\Console\Commands` namespace was replaced with `DragonCode\LaravelFeed\Commands`
