+++ tests/Feature/Console/Generation/NotFoundTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use DragonCode\LaravelFeed\Console\Commands\FeedGenerateCommand;
+use DragonCode\LaravelFeed\Commands\FeedGenerateCommand;
 use DragonCode\LaravelFeed\Exceptions\FeedNotFoundException;
 
 use function Pest\Laravel\artisan;
